Update and rename DDO69_background_subtraction.py to DDO69.py
This code, uses a UBV image of a galaxy from the LITTLE THINGS galaxy category, computes the background using the " photutils.background" library, then subtracts said background.
The output of this snippet is fed into a machine learning program, called "starnet++",  so that we can subtract all the stars and be left with only the galaxy. 
The starless image is then imported back into the code and used to create multiple plots such as a magnitude table.